Formatted for use with setup.py
I created a basic setup.py folder, and moved files into the folder,
'convnetpy'
