+++ convnetpy/layers/input.py
