+++ convnetpy/layers/nonlinearities.py
