+++ setup.py
@@ -1,0 +1,10 @@
+#!/usr/bin/env python
+
+from distutils.core import setup
+
+setup(name='Distutils',
+      version='1.0',
+      description='Python port of ConvNetJS',
+      url='https://github.com/Aaronduino/ConvNetPy',
+      packages=['convnetpy', 'convnetpy.layers'],
+     )
